+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+0.2.0
+-----
+- Removed following methods due to LastFM API changes:
+    - `track.ban`
+    - `track.get_buy_links`
+    - `track.get_fingerprint_metadata`
+    - `track.get_shouts`
+    - `track.get_top_fans`
+    - `track.share`
+    - `track.unban`
+    - `user.get_banned_tracks`
+    - `user.get_events`
+    - `user.get_neighbors`
+    - `user.get_neighbours`
+    - `user.get_new_releases`
+    - `user.get_past_events`
+    - `user.get_playlists`
+    - `user.get_recent_stations`
+    - `user.get_recommended_artists`
+    - `user.get_recommended_events`
+    - `user.get_shouts`
+    - `user.shout`
+
 0.1.1
 -----
 - Initial release